Add separate file for class definitions
