+++ src/content_mapblock.cpp
@@ -18,6 +18,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 */
 
 #include "content_mapblock.h"
+#include "content_mapblock_private.h"
 #include "util/numeric.h"
 #include "util/directiontables.h"
 #include "mapblock_mesh.h" // For MapBlock_LightColor() and MeshCollector
@@ -34,13 +35,6 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 // After this distance, it gives up and considers light level constant
 #define SMOOTH_LIGHTING_OVERSIZE 1.0
 
-struct LightFrame
-{
-	f32 lightsA[8];
-	f32 lightsB[8];
-	u8 light_source;
-};
-
 static const v3s16 light_dirs[8] = {
 	v3s16(-1, -1, -1),
 	v3s16(-1, -1,  1),
@@ -50,73 +44,6 @@ static const v3s16 light_dirs[8] = {
 	v3s16( 1, -1,  1),
 	v3s16( 1,  1, -1),
 	v3s16( 1,  1,  1),
-};
-
-class MapblockMeshGenerator
-{
-	MeshMakeData *data;
-	MeshCollector *collector;
-
-	INodeDefManager *nodedef;
-	scene::ISceneManager *smgr;
-	scene::IMeshManipulator *meshmanip;
-
-// options
-	bool enable_mesh_cache;
-
-// current node
-	v3s16 blockpos_nodes;
-	v3s16 p;
-	core::vector3df origin;
-	MapNode n;
-	const ContentFeatures *f;
-	u16 light;
-	LightFrame frame;
-
-// liquid-specific
-	bool top_is_same_liquid;
-	TileSpec tile_liquid;
-	TileSpec tile_liquid_top;
-	content_t c_flowing;
-	content_t c_source;
-	video::SColor color;
-	video::SColor color_liquid_top;
-	struct NeighborData {
-		f32 level;
-		content_t content;
-		bool is_same_liquid;
-		bool top_is_same_liquid;
-	};
-	NeighborData liquid_neighbors[3][3];
-	f32 corner_levels[2][2];
-
-	void prepareLiquidNodeDrawing(bool flowing);
-	void getLiquidNeighborhood(bool flowing);
-	void resetCornerLevels();
-	void calculateCornerLevels();
-	f32 getCornerLevel(u32 i, u32 k);
-	void drawLiquidSides(bool flowing);
-	void drawLiquidTop(bool flowing);
-
-// drawtypes
-	void drawLiquidNode(bool flowing);
-	void drawGlasslikeNode();
-	void drawGlasslikeFramedNode();
-	void drawAllfacesNode();
-	void drawTorchlikeNode();
-	void drawSignlikeNode();
-	void drawPlantlikeNode();
-	void drawFirelikeNode();
-	void drawFencelikeNode();
-	void drawRaillikeNode();
-	void drawNodeboxNode();
-	void drawMeshNode();
-
-	void drawNode();
-
-public:
-	MapblockMeshGenerator(MeshMakeData *input, MeshCollector *output);
-	void generate();
 };
 
 MapblockMeshGenerator::MapblockMeshGenerator(MeshMakeData *input, MeshCollector *output)
